requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pyyaml==6.0.2
qbittorrent-api==2024.3.60
anyio==4.4.0
python-multipart==0.0.9
httpx==0.27.2
